Fixed
64bit done and works windows.  Managed to reduce the amount of changes to a bare minimum and still needed some trusty long long's in places.

+++ WinQuake/d_polyse.c
@@ -129,7 +129,7 @@ void D_PolysetDraw (void)
 						// one extra because of cache line pretouching
 
 	a_spans = (spanpackage_t *)
-			(((long)&spans[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+			(((long long)&spans[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
 
 	if (r_affinetridesc.drawtype)
 	{
